move hooks to a separated file
